Move useGenerateDates to data folder
