+++ domains/rem/src/data/index.ts
@@ -1,5 +1,7 @@
 export { default as useFormState } from './useFormState';
 
+export { default as useGenerateDates } from './useGenerateDates';
+
 export { default as useInitialState } from './useInitialState';
 
 export { default as useFormStateManager } from './useFormStateManager';
