+++ domains/rem/src/ui/MultiStep/SubmitButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { __ } from '@eventespresso/i18n';
 
 import { Button, ButtonType, ButtonProps } from '@eventespresso/components';
-import { useGenerateDates } from '../generatedDates';
+import { useGenerateDates } from '../../data';
 
 const SubmitButton: React.FC<ButtonProps> = ({ onClick }) => {
 	// rDates and gDates, no exDates
