+++ domains/rem/src/ui/generatedDates/index.ts
@@ -1,4 +1,3 @@
 export { default as GeneratedDates } from './GeneratedDates';
-export { default as useGenerateDates } from './useGenerateDates';
 
 export * from './types';
